+++ kitty/graphics.c
@@ -2052,14 +2052,6 @@ id_range_filter_func(const ImageRef *ref UNUSED, Image *img, const void *data, C
 
 
 static bool
-number_filter_func(const ImageRef *ref, Image *img, const void *data, CellPixelSize cell UNUSED) {
-    const GraphicsCommand *g = data;
-    if (g->image_number && img->client_number == g->image_number) return !g->placement_id || ref->client_id == g->placement_id;
-    return false;
-}
-
-
-static bool
 x_filter_func(const ImageRef *ref, Image UNUSED *img, const void *data, CellPixelSize cell UNUSED) {
     if (ref->is_virtual_ref || is_cell_image(ref)) return false;
     const GraphicsCommand *g = data;
@@ -2098,9 +2090,8 @@ static void
 handle_delete_command(GraphicsManager *self, const GraphicsCommand *g, Cursor *c, bool *is_dirty, CellPixelSize cell) {
     if (self->currently_loading.loading_for.image_id) free_load_data(&self->currently_loading);
     GraphicsCommand d;
-    bool only_first_image = false;
     switch (g->delete_action) {
-#define I(u, data, func) filter_refs(self, data, g->delete_action == u, func, cell, only_first_image); *is_dirty = true; break
+#define I(u, data, func) filter_refs(self, data, g->delete_action == u, func, cell, false); *is_dirty = true; break
 #define D(l, u, data, func) case l: case u: I(u, data, func)
 #define G(l, u, func) D(l, u, g, func)
         case 0:
@@ -2117,9 +2108,18 @@ handle_delete_command(GraphicsManager *self, const GraphicsCommand *g, Cursor *c
             d.x_offset = c->x + 1; d.y_offset = c->y + 1;
             I('C', &d, point_filter_func);
         case 'n':
-        case 'N':
-            only_first_image = true;
-            I('N', g, number_filter_func);
+        case 'N': {
+            Image *img = img_by_client_number(self, g->image_number);
+            if (img) {
+                for (ref_map_itr ri = vt_first(&img->refs_by_internal_id); !vt_is_end(ri); ) { ImageRef *ref = ri.data->val;
+                    if (!g->placement_id || g->placement_id == ref->client_id) {
+                        ri = remove_ref_itr(img, ri);
+                        self->layers_dirty = true;
+                    } else ri = vt_next(ri);
+                }
+                if (!vt_size(&img->refs_by_internal_id) && (g->delete_action == 'N' || img->client_id == 0)) remove_image(self, img);
+            }
+        } break;
         case 'f':
         case 'F':
             if (handle_delete_frame_command(self, g, is_dirty) != NULL) {